reworks photo upload and database save callback to a utils function that can be used for each controller
